move logic from mainpage to todo page
