+++ src/app/components/mainpage/mainpage-routing.module.ts
@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { MainpageComponent } from './mainpage.component';
 import { ProfilePageComponent } from './profile-page/profile-page.component';
 import { TodoListPageComponent } from './todo-list-page/todo-list-page.component';
 import { WishListPageComponent } from './wish-list-page/wish-list-page.component';
