+++ src/app/components/mainpage/todo-list-page/todo-list-page.component.html
@@ -1,4 +1,4 @@
-<!-- <div class="body-container">
+<div class="body-container">
     <div class="cont" *ngIf="dataSource; else loading">
       <div class="cont-box" *ngIf="totalTodo; else notask">
         <div class="todo-box">
@@ -108,13 +108,7 @@
 
     <ng-template #loading>
       <div class="loading">
-        <mat-spinner></mat-spinner>
+        <!-- <mat-spinner></mat-spinner> -->
       </div>
     </ng-template>
-  </div> -->
-  <p>Todo list</p>
-  <p>Todo list</p>
-  <p>Todo list</p>
-  <p>Todo list</p>
-  <p>Todo list</p>
-  <p>Todo list</p>
+  </div>
