+++ src/app/components/mainpage/todo-list-page/todo-list-page.component.scss
@@ -1,0 +1,277 @@
+.page-container {
+    height: 100%;
+}
+
+.body-container {
+    display: flex;
+    width: 100%;
+    height: 100%;
+    z-index: 1;
+}
+
+.cont {
+    display: flex;
+    justify-content: center;
+    width: 100%;
+    margin-top: 8%;
+}
+
+.cont-box {
+    display: flex;
+    justify-content: center;
+    align-content: center;
+    width: 100%;
+
+    .todo-box {
+        width: 70%;
+
+        .form {
+            display: flex;
+            flex-direction: row;
+            justify-content: space-between;
+
+            .input-search {
+                width: 54%;
+            }
+
+            .input-category {
+                width: 26%;
+            }
+        }
+
+        .mat-form-field {
+            font-size: 14px;
+        }
+    }
+
+    .table-box {
+        box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.1);
+
+        table {
+            width: 100%;
+        }
+    }
+}
+
+.paginator {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+    background-color: var(--white);
+    padding-left: 18px;
+
+    .paginator-text {
+        font-size: 12px;
+    }
+
+    .progress {
+        width: 100%;
+    }
+}
+
+::ng-deep .form {
+    .mat-form-field {
+        .mat-form-field-label {
+            color: var(--black) !important;
+        }
+    }
+    
+    .mat-form-field-flex .mat-form-field-infix {
+        padding: 12px 0px !important;
+    }
+}
+
+.no-task-box {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin-top: 4%;
+    height: 100%;
+
+    .no-task-text {
+        margin-top: 16px;
+        font-size: 24px;
+    }
+}
+
+.loading {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    margin-top: 10%;
+    width: 100%;
+}
+
+.loading::ng-deep .mat-progress-spinner circle, .mat-spinner circle {
+    stroke: var(--white);
+}
+
+.check-col {
+    max-width: 60px;
+}
+
+.date-col {
+    max-width: 100px;
+}
+
+.tags-col {
+    max-width: 140px;
+}
+
+.category-col,
+.priority-col {
+    max-width: 90px;
+}
+
+.action-col {
+    display: flex;
+    justify-content: flex-end;
+    max-width: 140px;
+
+    .action-col-text {
+        padding-right: 12px;
+    }
+}
+
+.ready {
+    color: var(--medium-grey);
+}
+
+.unready {
+    color: var(--black);
+}
+
+.past {
+    color: var(--danger);
+}
+
+ul {
+    list-style-type: none;
+    margin: 0;
+    padding: 0;
+}
+
+.tags-box {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 4px;
+}
+
+.tag {
+    border-radius: 12px;
+    padding: 2px 8px;
+    font-size: 11px;
+}
+
+.ready-tag {
+    color: var(--white);
+    background-color: var(--light-grey);
+}
+
+.critical-tag {
+    color: var(--white);
+    background-color: var(--danger);
+}
+
+.high-tag {
+    color: var(--white);
+    background-color: var(--warning);
+}
+
+.medium-tag {
+    color: var(--white);
+    background-color: var(--success);
+}
+
+.low-tag {
+    color: var(--white);
+    background-color: var(--neutral);
+}
+
+.btn {
+    padding: 6px 12px;
+    margin-top: 3px;
+    margin-bottom: 26px;
+    color: var(--white);
+    background-color: var(--accent);
+    transition: all 0.4s ease-in-out;
+    cursor: pointer;
+
+    &:hover {
+        background-color: var(--accent-hover);
+    }
+}
+
+.btn-add {
+    padding: 6px 12px;
+    margin-top: 12px;
+    width: 140px;
+    color: var(--black);
+    background-color: var(--white);
+    transition: all 0.4s ease-in-out;
+    cursor: pointer;
+
+    &:hover {
+        color: var(--white);
+        background-color: var(--black);
+    }
+}
+
+.clear-btn {
+    color: var(--medium-grey);
+}
+
+.clear-btn:hover {
+    color: var(--grey);
+}
+
+.critical-icon {
+    color: var(--danger);
+}
+
+.high-icon {
+    color: var(--warning);
+}
+
+.medium-icon {
+    color: var(--success);
+}
+
+.low-icon {
+    color: var(--neutral);
+}
+
+.priority {
+    display: flex;
+
+    .fa-solid {
+        display: flex;
+        align-items: center;
+        font-size: 16px;
+        height: 20px;
+        width: 20px;
+    }
+}
+
+.p-priority {
+    font-size: 12px;
+}
+
+@media only screen and (max-width: 1400px) {
+    .cont {
+        margin-top: 9%;
+    }
+}
+
+@media only screen and (max-width: 990px) {
+    .cont {
+        margin-top: 11%;
+    }
+}
+
+@media only screen and (max-width: 620px) {
+    .cont {
+        margin-top: 14%;
+    }
+}
